requirements.txt

fastapi==0.75.0
redis-om==0.2.1
uvicorn==0.24.0.post1
requests
python-dotenv